Fixing Start/Close issues
