+++ pkg/reader/rmn_home.go
@@ -2,32 +2,21 @@ package reader
 
 import (
 	"context"
-	"crypto/ed25519"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"math/big"
-	"sync"
-	"time"
-
-	mapset "github.com/deckarep/golang-set/v2"
-	ragep2ptypes "github.com/smartcontractkit/libocr/ragep2p/types"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	"github.com/smartcontractkit/chainlink-common/pkg/services"
-	"github.com/smartcontractkit/chainlink-common/pkg/types"
-	"github.com/smartcontractkit/chainlink-common/pkg/types/query/primitives"
 
 	rmntypes "github.com/smartcontractkit/chainlink-ccip/commit/merkleroot/rmn/types"
-	"github.com/smartcontractkit/chainlink-ccip/pkg/consts"
 	"github.com/smartcontractkit/chainlink-ccip/pkg/contractreader"
-	"github.com/smartcontractkit/chainlink-ccip/pkg/logutil"
 	cciptypes "github.com/smartcontractkit/chainlink-ccip/pkg/types/ccipocr3"
 )
 
 const (
 	rmnMaxSizeCommittee = 256 // bitmap is 256 bits making the max committee size 256
-	MaxFailedPolls      = uint(10)
+	maxFailedPolls      = uint32(10)
 )
 
 type HomeNodeInfo = rmntypes.HomeNodeInfo
@@ -49,259 +38,55 @@ type RMNHome interface {
 	services.Service
 }
 
-var (
-	instances   = make(map[string]*rmnHomePoller)
-	instancesMu sync.Mutex
-)
-
-type rmnHomeState struct {
-	activeConfigDigest    cciptypes.Bytes32
-	candidateConfigDigest cciptypes.Bytes32
-	rmnHomeConfig         map[cciptypes.Bytes32]rmntypes.HomeConfig
+type rmnHome struct {
+	sync     services.StateMachine
+	bgPoller *rmnHomePoller
 }
 
-// rmnHomePoller polls the RMNHome contract for the latest RMNHomeConfigs
-// It is running in the background with a polling interval of pollingDuration
-type rmnHomePoller struct {
-	wg                   sync.WaitGroup
-	stopCh               services.StopChan
-	sync                 services.StateMachine
-	mutex                *sync.RWMutex
-	contractReader       contractreader.ContractReaderFacade
-	rmnHomeBoundContract types.BoundContract
-	lggr                 logger.Logger
-	rmnHomeState         rmnHomeState
-	failedPolls          uint
-	pollingDuration      time.Duration // How frequently the poller fetches the chain configs
-}
-
-// GetRMNHomePoller returns a rmnHomePoller instance if it already exists, else creates a new one.
-// Returned rmnHomePoller is always started, therefore user of that API doesn't need to care about
-// concurrency and accidentally starting the same poller multiple times.
+// NewRMNHome creates a new RMNHome and starts it immediately. RMNHome is very lightweight layer
+// on top of RMNHomePoller. Every consumer should follow the `Service` pattern in which they
+// are responsible for properly closing the service when done. (using Close() method)
 //
-// In most of the cases, there is going to be only a single RMNHome deployed on a single chain. However,
-// OCR3Config allows to pass different RMNHome addresses for different chains. Therefore, we need to
-// support that and maintain a separate rmnHomePoller for each RMNHome address.
-// Having a singleton here is aimed to reduce the background polling to the underlying RPC node.
-func GetRMNHomePoller(
+// RMNHome is a smart component that behind the scenes share the same RMNHomePoller instance.
+// Whenever all RMNHome instances are closed, the underlying RMNHomePoller instance is also closed.
+// As long as RMNHome remembers to Close() upon termination there should not be any orphaned poller instances
+// working in the background
+func NewRMNHome(
 	ctx context.Context,
 	lggr logger.Logger,
 	rmnHomeChainSelector cciptypes.ChainSelector,
 	rmnHomeAddress []byte,
 	contractReader contractreader.ContractReaderFacade,
-	pollingInterval time.Duration,
 ) (RMNHome, error) {
-	instancesMu.Lock()
-	defer instancesMu.Unlock()
-
-	hexEncodedAddr := " 0x" + hex.EncodeToString(rmnHomeAddress)
-	key := fmt.Sprintf("%s-%s", rmnHomeChainSelector.String(), hexEncodedAddr)
-
-	instance, ok := getInstance(key)
-	if ok {
-		lggr.Infow("RMNHomePoller already exists, reusing instance",
-			"chainSelector", rmnHomeChainSelector,
-			"address", hexEncodedAddr,
-		)
-		return instance, nil
-	}
-
-	rmnHomeBoundContract := types.BoundContract{
-		Address: hexEncodedAddr,
-		Name:    consts.ContractNameRMNHome,
-	}
-
-	if err := contractReader.Bind(ctx, []types.BoundContract{rmnHomeBoundContract}); err != nil {
-		return nil, fmt.Errorf("failed to bind RMNHome contract: %w", err)
-	}
-
-	rmnHomeReader := newRMNHomePoller(
+	bgPoller, err := getRMNHomePoller(
+		ctx,
+		lggr,
+		rmnHomeChainSelector,
+		rmnHomeAddress,
 		contractReader,
-		rmnHomeBoundContract,
-		logutil.WithComponent(lggr, "RMNHomePoller"),
-		pollingInterval,
+		HomeChainPollingInterval,
 	)
-
-	if err := rmnHomeReader.Start(ctx); err != nil {
-		return nil, fmt.Errorf("failed to start RMNHome reader: %w", err)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create RMNHomePoller: %w", err)
 	}
 
-	instances[key] = rmnHomeReader
-	return rmnHomeReader, nil
+	rmn := &rmnHome{bgPoller: bgPoller}
+
+	if err := rmn.Start(ctx); err != nil {
+		return nil, fmt.Errorf("failed to start RMNHome: %w", err)
+	}
+	return rmn, nil
 }
 
-func getInstance(key string) (*rmnHomePoller, bool) {
-	instance, ok := instances[key]
-	if !ok {
-		return nil, false
-	}
-
-	if err := instance.sync.Ready(); err != nil {
-		return nil, false
-	}
-
-	return instance, ok
-}
-
-func newRMNHomePoller(
-	contractReader contractreader.ContractReaderFacade,
-	rmnHomeBoundContract types.BoundContract,
-	lggr logger.Logger,
-	pollingInterval time.Duration,
-) *rmnHomePoller {
-	return &rmnHomePoller{
-		stopCh:               make(chan struct{}),
-		contractReader:       contractReader,
-		rmnHomeBoundContract: rmnHomeBoundContract,
-		rmnHomeState:         rmnHomeState{},
-		mutex:                &sync.RWMutex{},
-		failedPolls:          0,
-		lggr:                 lggr,
-		pollingDuration:      pollingInterval,
-	}
-}
-
-func (r *rmnHomePoller) Start(context.Context) error {
+func (r *rmnHome) Start(ctx context.Context) error {
 	return r.sync.StartOnce(r.Name(), func() error {
-		r.lggr.Infow("Start Polling RMNHome")
-		r.wg.Add(1)
-		go r.poll()
-		return nil
+		return r.bgPoller.Start(ctx, r)
 	})
 }
 
-func (r *rmnHomePoller) poll() {
-	defer r.wg.Done()
-	ctx, cancel := r.stopCh.NewCtx()
-	defer cancel()
-	// Initial fetch once poll is called before any ticks
-	if err := r.fetchAndSetRmnHomeConfigs(ctx); err != nil {
-		// Just log, don't return error as we want to keep polling
-		r.lggr.Errorw("Initial fetch of on-chain configs failed", "err", err)
-	}
-
-	ticker := time.NewTicker(r.pollingDuration)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			r.mutex.Lock()
-			r.failedPolls = 0
-			r.mutex.Unlock()
-			return
-		case <-ticker.C:
-			if err := r.fetchAndSetRmnHomeConfigs(ctx); err != nil {
-				r.mutex.Lock()
-				r.failedPolls++
-				r.mutex.Unlock()
-			} else {
-				r.mutex.Lock()
-				r.failedPolls = 0
-				r.mutex.Unlock()
-			}
-		}
-	}
-}
-
-func (r *rmnHomePoller) fetchAndSetRmnHomeConfigs(ctx context.Context) error {
-	var activeAndCandidateConfigs GetAllConfigsResponse
-	err := r.contractReader.GetLatestValue(
-		ctx,
-		r.rmnHomeBoundContract.ReadIdentifier(consts.MethodNameGetAllConfigs),
-		primitives.Unconfirmed,
-		map[string]interface{}{},
-		&activeAndCandidateConfigs,
-	)
-	if err != nil {
-		return fmt.Errorf("error fetching RMNHomeConfig: %w", err)
-	}
-	r.lggr.Infow("Fetched RMNHomeConfigs",
-		"activeConfig", activeAndCandidateConfigs.ActiveConfig.ConfigDigest,
-		"candidateConfig", activeAndCandidateConfigs.CandidateConfig.ConfigDigest,
-	)
-
-	if activeAndCandidateConfigs.ActiveConfig.ConfigDigest.IsEmpty() &&
-		activeAndCandidateConfigs.CandidateConfig.ConfigDigest.IsEmpty() {
-		return fmt.Errorf("both active and candidate config digests are empty")
-	}
-
-	r.setRMNHomeState(
-		activeAndCandidateConfigs.ActiveConfig.ConfigDigest,
-		activeAndCandidateConfigs.CandidateConfig.ConfigDigest,
-		convertOnChainConfigToRMNHomeChainConfig(
-			r.lggr,
-			activeAndCandidateConfigs.ActiveConfig,
-			activeAndCandidateConfigs.CandidateConfig,
-		),
-	)
-
-	return nil
-}
-
-func (r *rmnHomePoller) setRMNHomeState(
-	activeConfigDigest cciptypes.Bytes32,
-	candidateConfigDigest cciptypes.Bytes32,
-	rmnHomeConfig map[cciptypes.Bytes32]rmntypes.HomeConfig) {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-	s := &r.rmnHomeState
-
-	s.activeConfigDigest = activeConfigDigest
-	s.candidateConfigDigest = candidateConfigDigest
-	s.rmnHomeConfig = rmnHomeConfig
-}
-
-func (r *rmnHomePoller) GetRMNNodesInfo(configDigest cciptypes.Bytes32) ([]rmntypes.HomeNodeInfo, error) {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-	_, ok := r.rmnHomeState.rmnHomeConfig[configDigest]
-	if !ok {
-		return nil, fmt.Errorf("configDigest %s not found in RMNHomeConfig", configDigest)
-
-	}
-	return r.rmnHomeState.rmnHomeConfig[configDigest].Nodes, nil
-}
-
-func (r *rmnHomePoller) IsRMNHomeConfigDigestSet(configDigest cciptypes.Bytes32) bool {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-	_, ok := r.rmnHomeState.rmnHomeConfig[configDigest]
-	return ok
-}
-
-func (r *rmnHomePoller) GetFObserve(configDigest cciptypes.Bytes32) (map[cciptypes.ChainSelector]int, error) {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-	_, ok := r.rmnHomeState.rmnHomeConfig[configDigest]
-	if !ok {
-		return nil, fmt.Errorf("configDigest %s not found in RMNHomeConfig", configDigest)
-	}
-	return r.rmnHomeState.rmnHomeConfig[configDigest].SourceChainF, nil
-}
-
-func (r *rmnHomePoller) GetOffChainConfig(configDigest cciptypes.Bytes32) (cciptypes.Bytes, error) {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-	cfg, ok := r.rmnHomeState.rmnHomeConfig[configDigest]
-	if !ok {
-		return nil, fmt.Errorf("configDigest %s not found in RMNHomeConfig", configDigest)
-	}
-	return cfg.OffchainConfig, nil
-}
-
-func (r *rmnHomePoller) GetAllConfigDigests() (
-	activeConfigDigest cciptypes.Bytes32,
-	candidateConfigDigest cciptypes.Bytes32) {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-	return r.rmnHomeState.activeConfigDigest, r.rmnHomeState.candidateConfigDigest
-}
-
-func (r *rmnHomePoller) Close() error {
+func (r *rmnHome) Close() error {
 	err := r.sync.StopOnce(r.Name(), func() error {
-		defer r.wg.Wait()
-		close(r.stopCh)
-		return nil
+		return r.bgPoller.Close(r)
 	})
 
 	if errors.Is(err, services.ErrAlreadyStopped) {
@@ -311,121 +96,57 @@ func (r *rmnHomePoller) Close() error {
 	return err
 }
 
-func (r *rmnHomePoller) Ready() error {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
+func (r *rmnHome) GetRMNNodesInfo(configDigest cciptypes.Bytes32) ([]rmntypes.HomeNodeInfo, error) {
+	state := r.bgPoller.getRMNHomeState()
+	_, ok := state.rmnHomeConfig[configDigest]
+	if !ok {
+		return nil, fmt.Errorf("configDigest %s not found in RMNHomeConfig", configDigest)
+
+	}
+	return state.rmnHomeConfig[configDigest].Nodes, nil
+}
+
+func (r *rmnHome) IsRMNHomeConfigDigestSet(configDigest cciptypes.Bytes32) bool {
+	_, ok := r.bgPoller.getRMNHomeState().rmnHomeConfig[configDigest]
+	return ok
+}
+
+func (r *rmnHome) GetFObserve(configDigest cciptypes.Bytes32) (map[cciptypes.ChainSelector]int, error) {
+	state := r.bgPoller.getRMNHomeState()
+	_, ok := state.rmnHomeConfig[configDigest]
+	if !ok {
+		return nil, fmt.Errorf("configDigest %s not found in RMNHomeConfig", configDigest)
+	}
+	return state.rmnHomeConfig[configDigest].SourceChainF, nil
+}
+
+func (r *rmnHome) GetOffChainConfig(configDigest cciptypes.Bytes32) (cciptypes.Bytes, error) {
+	state := r.bgPoller.getRMNHomeState()
+	cfg, ok := state.rmnHomeConfig[configDigest]
+	if !ok {
+		return nil, fmt.Errorf("configDigest %s not found in RMNHomeConfig", configDigest)
+	}
+	return cfg.OffchainConfig, nil
+}
+
+func (r *rmnHome) GetAllConfigDigests() (
+	activeConfigDigest cciptypes.Bytes32,
+	candidateConfigDigest cciptypes.Bytes32,
+) {
+	state := r.bgPoller.getRMNHomeState()
+	return state.activeConfigDigest, state.candidateConfigDigest
+}
+
+func (r *rmnHome) Ready() error {
 	return r.sync.Ready()
 }
 
-func (r *rmnHomePoller) HealthReport() map[string]error {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-	if r.failedPolls >= MaxFailedPolls {
-		err := fmt.Errorf("polling failed %d times in a row (maximum allowed: %d)", r.failedPolls, MaxFailedPolls)
-		r.sync.SvcErrBuffer.Append(err)
-	}
-	return map[string]error{r.Name(): r.sync.Healthy()}
+func (r *rmnHome) HealthReport() map[string]error {
+	return r.bgPoller.HealthReport()
 }
 
-func (r *rmnHomePoller) Name() string {
-	return r.lggr.Name()
-}
-
-func validate(config VersionedConfig) error {
-	// check if the versionesconfigwithdigests are set (can be empty)
-	if config.ConfigDigest.IsEmpty() {
-		return fmt.Errorf("configDigest is empty")
-	}
-	return nil
-}
-
-func convertOnChainConfigToRMNHomeChainConfig(
-	lggr logger.Logger,
-	primaryConfig VersionedConfig,
-	secondaryConfig VersionedConfig,
-) map[cciptypes.Bytes32]rmntypes.HomeConfig {
-	if primaryConfig.ConfigDigest.IsEmpty() && secondaryConfig.ConfigDigest.IsEmpty() {
-		lggr.Warnw("no on chain RMNHomeConfigs found, both digests are empty")
-		return map[cciptypes.Bytes32]rmntypes.HomeConfig{}
-	}
-
-	versionedConfigWithDigests := []VersionedConfig{primaryConfig}
-	if !secondaryConfig.ConfigDigest.IsEmpty() {
-		versionedConfigWithDigests = append(versionedConfigWithDigests, secondaryConfig)
-	}
-
-	rmnHomeConfigs := make(map[cciptypes.Bytes32]rmntypes.HomeConfig)
-	for _, versionedConfig := range versionedConfigWithDigests {
-		err := validate(versionedConfig)
-		if err != nil {
-			lggr.Warnw("invalid on chain RMNHomeConfig", "err", err)
-			continue
-		}
-
-		nodes := make([]rmntypes.HomeNodeInfo, len(versionedConfig.StaticConfig.Nodes))
-		for i, node := range versionedConfig.StaticConfig.Nodes {
-			pubKey := ed25519.PublicKey(node.OffchainPublicKey[:])
-
-			nodes[i] = rmntypes.HomeNodeInfo{
-				ID:                    rmntypes.NodeID(i),
-				PeerID:                ragep2ptypes.PeerID(node.PeerID),
-				OffchainPublicKey:     &pubKey,
-				SupportedSourceChains: mapset.NewSet[cciptypes.ChainSelector](),
-				StreamNamePrefix:      "ccip-rmn/v1_6/", // todo: when contract is updated, this should be fetched from the contract
-			}
-		}
-
-		homeFMap := make(map[cciptypes.ChainSelector]int)
-
-		for _, chain := range versionedConfig.DynamicConfig.SourceChains {
-			homeFMap[chain.ChainSelector] = int(chain.FObserve)
-			for j := 0; j < len(nodes); j++ {
-				isObserver, err := IsNodeObserver(chain, j, len(nodes))
-				if err != nil {
-					lggr.Warnw("failed to check if node is observer", "err", err)
-					continue
-				}
-				if isObserver {
-					nodes[j].SupportedSourceChains.Add(chain.ChainSelector)
-				}
-			}
-		}
-
-		rmnHomeConfigs[versionedConfig.ConfigDigest] = rmntypes.HomeConfig{
-			Nodes:          nodes,
-			SourceChainF:   homeFMap,
-			ConfigDigest:   versionedConfig.ConfigDigest,
-			OffchainConfig: versionedConfig.DynamicConfig.OffchainConfig,
-		}
-	}
-	return rmnHomeConfigs
-}
-
-// IsNodeObserver checks if a node is an observer for the given source chain
-func IsNodeObserver(sourceChain SourceChain, nodeIndex int, totalNodes int) (bool, error) {
-	if totalNodes > rmnMaxSizeCommittee || totalNodes <= 0 {
-		return false, fmt.Errorf("invalid total nodes: %d", totalNodes)
-	}
-
-	if nodeIndex < 0 || nodeIndex >= totalNodes {
-		return false, fmt.Errorf("invalid node index: %d", nodeIndex)
-	}
-
-	// Validate the bitmap
-	maxValidBitmap := new(big.Int).Lsh(big.NewInt(1), uint(totalNodes))
-	maxValidBitmap.Sub(maxValidBitmap, big.NewInt(1))
-	if sourceChain.ObserverNodesBitmap.Cmp(maxValidBitmap) > 0 {
-		return false, fmt.Errorf("invalid observer nodes bitmap")
-	}
-
-	// Create a big.Int with 1 shifted left by nodeIndex
-	mask := new(big.Int).Lsh(big.NewInt(1), uint(nodeIndex))
-
-	// Perform the bitwise AND operation
-	result := new(big.Int).And(sourceChain.ObserverNodesBitmap, mask)
-
-	// Check if the result equals the mask
-	return result.Cmp(mask) == 0, nil
+func (r *rmnHome) Name() string {
+	return "RMNHome"
 }
 
 // GetAllConfigsResponse mirrors RMNHome.sol's getAllConfigs() return value.
@@ -466,4 +187,4 @@ type SourceChain struct {
 	ObserverNodesBitmap *big.Int                `json:"observerNodesBitmap"`
 }
 
-var _ RMNHome = (*rmnHomePoller)(nil)
+var _ RMNHome = (*rmnHome)(nil)
